+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to semantic versioning.
 - Add `DOM` module.
 - Add `converge`, `invoke`, and `invokeOn` functions to the `Function` module.
 - Add `whenInvocationCount` and `execute` functions to the `IO` module.
+- Add `execute` function to the `Task` module.
 - Add `isFinite` and `toFinite` functions to the `Number` module.
 - Add `noneAs`, `invert`, and `toMonoid` functions to the `Option` module.
 - Duplicate `filterA` from fp-ts-contrib into the `Array` and `ReadonlyArray` modules.